feat: Add support for parsing JSON request bodies
This commit adds the `body-parser` middleware to the application to support parsing JSON request bodies. This will allow clients to send JSON data in the request body, which can be useful for creating or updating resources.

The `body-parser` middleware is configured to parse JSON data by default, so no additional configuration is necessary.

This commit also updates the `tasksRoute.ts` file to use the `TaskController` class instead of defining the route handlers inline.

Finally, this commit adds a `README.md` file to the project with some basic documentation on how to run the application and use the API.