chore: Update tsconfig.json
This commit updates the `tsconfig.json` file to include the `baseUrl` and `paths` compiler options. These options are used to configure module resolution in the application.

The `baseUrl` option is set to `./src`, which tells TypeScript to resolve module imports relative to the `src` directory.

The `paths` option is used to define aliases for module imports. In this case, we have defined an alias for the `@controllers` directory, which contains the controller classes for the application.

This commit also removes the `skipLibCheck` option from the `compilerOptions` object. This option was causing some issues with type checking in the application.