feat: Add Task model
This commit adds a new model to the project called `Task`. The `Task` model extends the `Model` class from the `sequelize` library and defines the schema for the `tasks` table in the database.

The `Task` model has five properties: `id`, `title`, `description`, `completed`, and `date`. The `id` property is an auto-incrementing integer and serves as the primary key for the table. The `title` and `description` properties are strings that represent the title and description of the task. The `completed` property is a boolean that indicates whether the task has been completed or not. The `date` property is a date that represents the date the task was created.

This commit also updates the `sequelize` instance to use the `Task` model when defining the `tasks` table in the database.