feat: Add getAllTasks method to TaskService
This commit adds a new method to the `TaskService` class called `getAllTasks`. This method retrieves all tasks from the database using the `Task.findAll` method from the `sequelize` library.

The `getAllTasks` method returns a Promise that resolves to an array of `Task` objects. If there are no tasks in the database, the Promise resolves to an empty array.

This commit also updates the `TaskController` class to use the `getAllTasks` method to retrieve all tasks from the database.