feat: Add getAllTasks route to TasksController
This commit adds a new route to the `TasksController` class called `getAllTasks`. This route handles GET requests to the `/tasks` endpoint and retrieves all tasks from the database using the `getAllTasks` method from the `TasksService` class.

The `getAllTasks` route returns a JSON response that contains an array of `Task` objects. If there are no tasks in the database, the response contains an empty array.

This commit also updates the `index.ts` file to use the `TasksController` class and register the `getAllTasks` route.